add AuthContextProvider component
