+++ src/App.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import {Route, Routes } from 'react-router-dom';
 import { About } from './pages/About';
 import { Detail } from './components/Detail';
@@ -6,23 +6,24 @@ import { Home } from './pages/Home';
 import { Login } from './components/Login';
 import { Layout } from './components/Layout';
 import { Signup } from './components/Signup';
+import { AuthContextProvider } from './context/AuthContextProvider';
 
 function App() {
-	const tokenStorage = localStorage.getItem('token') 
-	const [token, setToken] = useState(tokenStorage);
 
 	return (
 		<>
+		<AuthContextProvider>
 			<Routes>
 				<Route index element={<Login/>}/>	
 				<Route path='signup' element={<Signup/>}/>
-				<Route path='login' element={<Login setToken={setToken}/>}/>	
-				<Route element={<Layout token={token} setToken={setToken} redirectPath ='/login'/>}>
+				<Route path='login' element={<Login/>}/>			
+				<Route element={<Layout redirectPath ='/login'/>}>
 					<Route path='home' element={<Home/>}/>
 					<Route path='about' element={<About/>}/>
 					<Route path='detail' element={<Detail/>}/>	
 				</Route>
 			</Routes>
+		</AuthContextProvider>
 		</>
 	);
 }
